added 2 methods to Support + other minor changes.
ApplicationWindow.setIconByResourceName: The signature now forces the
the passed string to be final.

Support.displayException: added an additional bit of text to the end of
error messages clarifying whether it is a fatal error resulting in the
termination of the application or a recoverable error thus allowing you
to continue.

Support.getResourceByName: The signature now forces the the passed
string to be final.

Support.openWebPageInDefaultBrowser: this is a new method that does
exactly what it says: opens a given URL in a new tab/window with the
user's default web browser. Obviously requires an internet connection,
otherwise it will generate an exception.

Support.playAudioClipFromURL: another new self explanatory method: you
give it the URL to a sound file such as a WAV audio file and it will
play it. Obviously requires an internet connection, otherwise it will
generate an exception.